修改心跳机制

+++ eagle-ui/src/main/java/eagle/jfaster/org/pojo/ClientServiceInfo.java
@@ -78,7 +78,7 @@ public class ClientServiceInfo implements ServiceCommonSetter {
 
     @Setter
     @Getter
-    Long heartbeat ;
+    Integer heartbeat ;
 
     @Setter
     @Getter
